replaced fps class with kairos timing library. This brings a number of new time based features.

+++ BaseProject/GameEngine/Game/CustomClasses/Rect.hpp
@@ -13,12 +13,12 @@
 
 struct myRect
 {
-    myRect(int l, int t, int r, int b) : left{l},right{r},top{t},base{b},width{r-l},height{b-t},area{width * height} {};
-    myRect(sf::IntRect ir)             : left{ir.left},right{ir.left + ir.width},top{ir.top},base{ir.top + ir.height},width{ir.width},height{ir.height},area{width * height}{};
-    myRect(int w, int h)               : left{0},right{w},top{0},base{h},width{w},height{h},area{width * height}{};
+    myRect(float l, float t, float r, float b) : left{l},right{r},top{t},base{b},width{r-l},height{b-t},area{width * height} {};
+    myRect(sf::IntRect ir)             : left{(float)ir.left},right{(float)ir.left + (float)ir.width},top{(float)ir.top},base{(float)ir.top + (float)ir.height},width{(float)ir.width},height{(float)ir.height},area{width * height}{};
+    myRect(float w, float h)               : left{0},right{w},top{0},base{h},width{w},height{h},area{width * height}{};
     myRect()                           : left{0},right{1},top{0},base{1},width{1},height{1},area{width*height} {};
     
-    void move(int x, int y)
+    void move(float x, float y)
     {
         left  += x;
         right += x;
@@ -26,12 +26,12 @@ struct myRect
         base  += y;
     };
     
-    void SetPosition(int x, int y)
+    void SetPosition(float x, float y)
     {
         move(x - left, y - top);
     }
     
-    int left{0},right{0},top{0},base{0},width{0},height{0},area{0};
+    float left{0},right{0},top{0},base{0},width{0},height{0},area{0};
     
 };
